requirements.txt
aiogram==3.0.0b6
wikipedia-api==0.7.1
python-dotenv==1.0.0
aiohttp==3.8.5
requests==2.32.3
charset-normalizer==3.4.0
idna==3.10
certifi==2024.8.30
urllib3==2.2.3
